backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aniso8601==10.0.0
attrs==24.3.0
Authlib==1.3.2
bcrypt==4.2.1
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloud-sql-python-connector==1.2.4
cryptography==44.0.0
ecdsa==0.19.0
Flask==3.0.0
Flask-Cors==5.0.0
flask-restplus==0.13.0
flask-restx==1.3.0
frozenlist==1.5.0
google-api-core==2.23.0
google-auth==2.37.0
google-cloud-core==2.4.1
google-cloud-datastore==2.20.1
google-cloud-secret-manager==2.23.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.14.0
grpcio==1.68.0
grpcio-status==1.68.0
gunicorn==22.0.0
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
multidict==6.1.0
packaging==24.2
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.22.3
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.37
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
yarl==1.18.3
